automation/requirements_auth.txt
#!/usr/bin/env python3
# requirements.txt for MIA.vn Authentication Service

# Core web framework
Flask==2.3.3
Flask-CORS==4.0.0

# Google Sheets integration
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.2

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
